+++ src/main/java/kr/co/inntavern/dripking/dto/Response/DistilleryResponseDTO.java
@@ -17,9 +17,9 @@ public class DistilleryResponseDTO {
 
     private String img_url;
 
-    private float latitude;//위도
+    private Double latitude;//위도
 
-    private float longitude;//경도
+    private Double longitude;//경도
 
     private Long destination_id;
