+++ src/main/java/kr/co/inntavern/dripking/model/Distillery.java
@@ -3,8 +3,6 @@ package kr.co.inntavern.dripking.model;
 import jakarta.persistence.*;
 import lombok.*;
 
-import java.util.List;
-
 @Setter
 @Getter
 @Builder
@@ -27,10 +25,10 @@ public class Distillery extends Item{
     private String description;
 
     @Column
-    private float latitude;//위도
+    private Double latitude;//위도
 
     @Column
-    private float longitude;//경도
+    private Double longitude;//경도
 
     @Column(nullable = false)
     private String img_url;
